Renamed btcutil/btcec to btcutil/koblitz.
