+++ btcutil/address.go
@@ -8,10 +8,9 @@ import (
 	"bytes"
 	"encoding/hex"
 	"errors"
-
 	"github.com/mit-dci/lit/btcutil/base58"
-	"github.com/mit-dci/lit/btcutil/btcec"
 	"github.com/mit-dci/lit/btcutil/chaincfg"
+	"github.com/mit-dci/lit/btcutil/koblitz"
 	"github.com/mit-dci/lit/crypto/ripemd160"
 )
