+++ btcutil/blockchain/compress.go
@@ -5,7 +5,7 @@
 package blockchain
 
 import (
-	"github.com/mit-dci/lit/btcutil/btcec"
+	"github.com/mit-dci/lit/btcutil/koblitz"
 	"github.com/mit-dci/lit/btcutil/txscript"
 )
