+++ btcutil/koblitz/ciphering.go
