+++ btcutil/koblitz/doc.go
