+++ btcutil/koblitz/field.go
