+++ btcutil/koblitz/precompute.go
