+++ btcutil/koblitz/privkey.go
