+++ lnp2p/util.go
@@ -2,7 +2,7 @@ package lnp2p
 
 import (
 	"github.com/mit-dci/lit/bech32"
-	"github.com/mit-dci/lit/btcutil/btcec"
+	"github.com/mit-dci/lit/btcutil/koblitz"
 	"github.com/mit-dci/lit/crypto/fastsha256"
 	"github.com/mit-dci/lit/lnio"
 	"strings"
