+++ lnutil/curvelib.go
@@ -6,7 +6,7 @@ import (
 	"math/big"
 	"sort"
 
-	"github.com/mit-dci/lit/btcutil/btcec"
+	"github.com/mit-dci/lit/btcutil/koblitz"
 	"github.com/mit-dci/lit/btcutil/chaincfg/chainhash"
 )
