+++ lnutil/tracker.go
@@ -6,15 +6,13 @@ import (
 	"encoding/hex"
 	"encoding/json"
 	"errors"
+	"github.com/mit-dci/lit/btcutil/koblitz"
+	"github.com/mit-dci/lit/logging"
+	"golang.org/x/net/proxy"
 	"net/http"
 	"net/url"
 	"strings"
 	"time"
-
-	"github.com/mit-dci/lit/logging"
-
-	"github.com/mit-dci/lit/btcutil/btcec"
-	"golang.org/x/net/proxy"
 )
 
 type announcement struct {
