+++ qln/dualfund.go
@@ -2,7 +2,7 @@ package qln
 
 import (
 	"fmt"
-	"github.com/mit-dci/lit/btcutil/btcec"
+	"github.com/mit-dci/lit/btcutil/koblitz"
 	"github.com/mit-dci/lit/btcutil/txsort"
 	"github.com/mit-dci/lit/consts"
 	"github.com/mit-dci/lit/elkrem"
