+++ qln/netio.go
@@ -2,7 +2,7 @@ package qln
 
 import (
 	"fmt"
-	"github.com/mit-dci/lit/btcutil/btcec"
+	"github.com/mit-dci/lit/btcutil/koblitz"
 	"github.com/mit-dci/lit/lnio"
 	"github.com/mit-dci/lit/lnutil"
 	"github.com/mit-dci/lit/logging"
