+++ wallit/basewalletif.go
@@ -2,17 +2,15 @@ package wallit
 
 import (
 	"fmt"
-	"sort"
-
-	"github.com/mit-dci/lit/logging"
-
-	"github.com/mit-dci/lit/btcutil/btcec"
 	"github.com/mit-dci/lit/btcutil/chaincfg/chainhash"
+	"github.com/mit-dci/lit/btcutil/koblitz"
 	"github.com/mit-dci/lit/coinparam"
 	"github.com/mit-dci/lit/lnutil"
+	"github.com/mit-dci/lit/logging"
 	"github.com/mit-dci/lit/portxo"
 	"github.com/mit-dci/lit/uspv"
 	"github.com/mit-dci/lit/wire"
+	"sort"
 )
 
 /*
